AURN API (WIP)
As there is no official AURN API for python, some bodging has occured. get_metadata function scrapes the AURN website for a) the csv containing info on all sites and b) the download code for all data (Usually 3 or 4 characters). This info can be used to get all metadata for the site as well as send a http request to download the measurements as a csv (TODO).

The scraping occurs via the use of lxml, which searches the HTML source using XPath strings. If the website changes, hopefully the only modification needed is to the XPaths. All URLs are also included in config in case future modification is needed