Documentation and format change
Full documentation added for the csv download method. The new csv format was also changed to add a unit column as theoretically the instrument could change midway through a year which should be represented in the data somehow. This means influx will have to store unit as a tag instead of in the measurement name, no harm no foul